# DB query optimization and reducing sqlalchemy logs