--- a/pebblo/app/service/local_ui/loader_apps.py
+++ b/pebblo/app/service/local_ui/loader_apps.py
@@ -3,7 +3,7 @@ from typing import List, Tuple
 
 from fastapi import status
 from sqlalchemy.ext.declarative import declarative_base
-
+from pebblo.app.utils.utils import timeit
 from pebblo.app.enums.enums import CacheDir, ReportConstants
 from pebblo.app.models.db_models import (
     DataSource,
@@ -58,13 +58,15 @@ class LoaderApp:
         This function finds snippet details based on labels
         """
         response = []
-        for snippet_id in snippet_ids:
+        result, output = self.db.query_by_list(AiSnippetsTable, "id", snippet_ids)
+
+        for row in output:
             if len(response) >= ReportConstants.SNIPPET_LIMIT.value:
                 break
-            result, output = self.db.query(AiSnippetsTable, {"id": snippet_id})
+
             if not result or len(output) == 0:
                 continue
-            snippet_details = output[0].data
+            snippet_details = row.data
             entity_details = {}
             topic_details = {}
             if snippet_details.get("topicDetails") and snippet_details[
@@ -351,6 +353,7 @@ class LoaderApp:
         )
         return loader_response
 
+    @timeit
     def get_all_loader_apps(self):
         """
         Returns all necessary loader app details required for get all app functionality.
@@ -402,6 +405,7 @@ class LoaderApp:
             # Closing the session
             self.db.session.close()
 
+    @timeit
     def get_loader_app_details(self, db: SQLiteClient, app_name: str) -> str:
         """
         This function is being used by the loader_doc_service to get data needed to generate pdf.