--- a/pebblo/app/storage/sqlite_db.py
+++ b/pebblo/app/storage/sqlite_db.py
@@ -1,8 +1,10 @@
 import logging
+from typing import List, Type
 
 from sqlalchemy import and_, create_engine, func, text
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.orm.attributes import flag_modified
+from sqlalchemy.orm.decl_api import DeclarativeMeta
 
 from pebblo.app.enums.enums import CacheDir
 from pebblo.app.storage.database import Database
@@ -110,17 +112,28 @@ class SQLiteClient(Database):
             return False, err
 
     @timeit
-    def query_by_list(self, table_obj, key, ids):
+    def query_by_list(
+        self,
+        table_obj: Type[DeclarativeMeta],
+        filter_key: str,
+        filter_values: List[str],
+    ):
         """
         Pass filter like list. For example get snippets with ids in [<id1>, <id2>]
+        :param table_obj: Table object on which query is to be performed
+        :param filter_key: Search key
+        :param filter_values: List of strings to be added to filter criteria
         """
-        # This function is not in use right now, But in the local_ui it will get used.
         table_name = table_obj.__tablename__
         try:
             logger.debug(f"Fetching data from table {table_name}")
             output = (
                 self.session.query(table_obj)
-                .filter(func.json_extract(table_obj.data, f"$.{key}").in_(ids))
+                .filter(
+                    func.json_extract(table_obj.data, f"$.{filter_key}").in_(
+                        filter_values
+                    )
+                )
                 .all()
             )
             return True, output