--- a/tests/app/storage/test_sqlite_db.py
+++ b/tests/app/storage/test_sqlite_db.py
@@ -33,70 +33,65 @@ def test_query_by_list_success(sqlite_client, mocker):
     mock_filter = mock_query.filter.return_value
     mock_filter.all.return_value = ["result1", "result2"]  # Mocked results
 
-    # Mock func.json_extract
-    mock_json_extract = mocker.patch("sqlalchemy.sql.func.json_extract")
-
     # Call the method
     success, result = sqlite_client.query_by_list(table_obj, filter_key, filter_values)
-
-    # Ensure that `json_extract` was called with the correct arguments
-    mock_json_extract.assert_called_once_with(table_obj.data, f"$.{filter_key}")
 
     # Assertions
     assert success is True
     assert result == ["result1", "result2"]
 
+    # Ensure the query was called only once (no pagination)
+    assert mock_session.query().filter().all.call_count == 1
 
-def test_query_by_list_max_filter_values(sqlite_client, mocker):
+
+def test_query_by_list_page_size(sqlite_client):
     """Test successful query with query_by_list to verify max_filter_limit"""
     mock_session = sqlite_client.session
     table_obj = AiSnippetsTable
     filter_key = "id"
-    filter_values = ["snippet_id1", "snippet_id2", "snippet_id3", "snippet_id4"]
-    max_filter_values = 2
+    filter_values = ["snippet_id1", "snippet_id2", "snippet_id3", "snippet_id4", "snippet_id5"]
+    page_size = 2
 
     # Mocking query result
-    mock_query = mock_session.query.return_value
-    mock_filter = mock_query.filter.return_value
-    mock_filter.all.return_value = ["result1", "result2"]  # Mocked results
-
-    # Mock func.json_extract
-    mock_json_extract = mocker.patch("sqlalchemy.sql.func.json_extract")
+    mock_result_page_1 = ["result1", "result2"]
+    mock_result_page_2 = ["result3", "result4"]
+    mock_result_page_3 = ["result5"]
+    mock_query = mock_session.query().filter().all
+    mock_query.side_effect = [mock_result_page_1, mock_result_page_2, mock_result_page_3]
 
     # Call the method
     success, result = sqlite_client.query_by_list(
-        table_obj, filter_key, filter_values, max_filter_values
+        table_obj, filter_key, filter_values, page_size
     )
-
-    # Ensure that `json_extract` was called with the correct arguments
-    mock_json_extract.assert_called_once_with(table_obj.data, f"$.{filter_key}")
 
     # Assertions
     assert success is True
     assert result == [
         "result1",
         "result2",
-    ]  # this should only limit first two items and skip rest two in the input.
+        "result3",
+        "result4",
+        "result5"
+    ]
 
 
-def test_query_by_list_failure(sqlite_client, mocker):
-    """Test query_by_list failure due to an exception."""
+def test_query_by_list_failure(sqlite_client):
     mock_session = sqlite_client.session
-    mock_table_obj = AiDataSourceTable
-    filter_key = "invalid_key"
+    mock_table_obj = AiSnippetsTable
+    filter_key = "id"
     filter_values = ["value1", "value2"]
 
-    # Mocking an exception when calling the query
-    mock_session.query.side_effect = Exception(
-        "DB Error. Key with name 'invalid_key' does not exists."
+    # Create mock data
+    page_size = "abcd"  # invalid page size
+
+    # Call the query_by_list function
+    success, results = sqlite_client.query_by_list(
+        table_obj=mock_table_obj,
+        filter_key=filter_key,
+        filter_values=filter_values,
+        page_size=page_size
     )
 
-    # Call the method
-    success, result = sqlite_client.query_by_list(
-        mock_table_obj, filter_key, filter_values
-    )
-
-    # Assertions
     assert success is False
-    assert result == []
-    mock_session.query.assert_called_once_with(mock_table_obj)
+    assert results == []
+    mock_session.query.assert_not_called()